Refactored landing page code
