+++ src/pages/layout/ActionButton.tsx
@@ -1,0 +1,28 @@
+import React from "react";
+
+interface ActionButtonProps {
+  href: string;
+  type?: "primary" | "secondary";
+  target?: string;
+  children: React.ReactNode;
+  buttonClassName?: string;
+}
+
+export function ActionButton({
+  href,
+  type = "primary",
+  target,
+  children,
+  buttonClassName,
+}: ActionButtonProps) {
+  return (
+    <a
+      className={`ActionButton ${type} ${buttonClassName}`}
+      rel="noopener"
+      href={href}
+      target={target}
+    >
+      {children}
+    </a>
+  );
+}
