+++ src/pages/layout/Heading.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+
+interface HeadingProps {
+  text: string;
+}
+
+export const Heading: React.FC<HeadingProps> = ({ text }) => {
+  return <h2 className="Heading">{text}</h2>;
+};
+
