+++ src/pages/utils/textcontent.js
@@ -1,0 +1,93 @@
+
+export const textContent = {
+    intro: `
+    <ul>
+  <li>Talawa is an open-source software project to help manage organizations like yours </li>
+  <li>Your community becomes closer using our mobile app with social media features </li>
+  <li>The Admin Web Portal makes member, event and volunteer management simpler.</li>
+  </ul>
+    `,
+    nativeCode: `
+    <ul>
+  <li>Easily schedule events and track attendance. Reports show who’s coming, and how often </li>
+  <li>Stay in contact with your membership with individual or group emails, now and in the future </li>
+  <li>Track your volunteer time and their activity tasks. Make your planning a success </li>
+  <li>Our intuitive design helps streamline work processes and increases adoption so everyone can benefit </li>
+  </ul>
+    `,
+    codeExample: `
+  import React from 'react';
+  import {Header} from './Header';
+  
+  const WelcomeScreen = () => (
+    <View>
+      <Header title="Welcome to Talawa"/>
+      <Text style={heading}>Step One</Text>
+      <Text>
+      Talawa is powerful, yet intuitive community management software made for religious
+      organizations, clubs, charities and other non-profits
+      </Text>
+      <Text>
+      Designed to help you save time, reduce frustration, and accomplish your goals.
+      </Text>
+      <Text>
+      Giving, event management, social media features, plus more.
+      </Text>
+      <Text>
+        Read the docs to discover what to do next:
+      </Text>
+  
+     </View>
+  );
+    `,
+    forEveryone: `
+    <ul>
+    <li>Simple to use member dashboards</li>
+    <li>Member notifications</li>
+    <li>Event and volunteer calendars</li>
+    <li>Donation management </li>
+    <li>Advertising from local companies </li>
+    <li>Multi-organization support </li>
+    <li>Plugin support for cloud service providers</li>
+    </ul>
+   `,
+    crossPlatform: `
+    <ul>
+  <li>View the social media news feed </li>
+  <li>Share posts, images and videos </li>
+  <li>Create events for your groups </li>
+  <li>Group chat with other members </li>
+  <li>Receive the latest community news </li>
+  <li>Get reminders and notifications </li>
+  <li>Make your community stronger </li>
+  </ul>
+    `,
+    sixthPanel: `Talawa is an open-source project. Our active community of volunteers is constantly adding features to make administration easier.</br></br>Their varied perspectives has created a product that is easily adaptable to the cultures and languages of their countries. Contributions from 19 different timezones.
+    `,
+    seventhPanel:`Talawa was created by The Palisadoes Foundation in 2019 as part of its social outreach programs in Jamaica. Our work was sponsored by local companies wanting to create a globally competitive workforce on the island by sponsoring philanthropic software development internships for university students.<br/><br/>
+    In 2021, Palisadoes was accepted into the prestigious Google Summer of Code (GSoC) which expanded  the development team to volunteers around the world. That same year we participated in the GitHub India, GitHub Externship program for both the winter and summer cohorts.<br/><br/>
+    Palisadoes was also selected for GSoC in 2022 and 2023. In 2022 the Google Season of Docs (GsoD) selected Palisadoes as a participant. In 2023 our Talawa projects were included in the Outreachy open-source internship program.
+    `,
+    eightPanel:`<p>Our developer community members are talented, passionate, and amazing. They are always suggesting, coding and documenting new ideas for our software.</p>
+    <p>They span across the globe. Explore this space to learn something new from the community.</p>
+    <br/>
+    <ul>
+      <li><a href="/docs/">Explore this site</a></li>
+      <li><a href="https://github.com/PalisadoesFoundation" target="_self">Visit our GitHub pages</a></li>
+      <li><a href="https://github.com/PalisadoesFoundation/talawa#readme" target="_self">Join our Slack channel</a></li>
+    </ul>
+    <br/>
+    <strong>Jump in and get started!</strong>
+    
+    `,
+    fastRefresh: `
+  Talawa was created by The Palisadoes Foundation in 2019 as part of its social outreach programs in Jamaica. Our work was sponsored by local companies wanting to create a globally competitive workforce on the island by sponsoring philanthropic software development internships for university students.<br/><br/>In 2021, Palisadoes was accepted into the prestigious Google Summer of Code (GSoC) which expanded  the development team to volunteers around the world. That same year we participated in the GitHub India, GitHub Externship program for both the winter and summer cohorts. <br/><br/>4.	Palisadoes was also selected for GSoC in 2022 and 2023. In 2022 the Google Season of Docs (GsoD) selected Palisadoes as a participant. In 2023 our Talawa projects were included in the Outreachy open source internship program.
+  
+    `,
+    talks: `
+    Get to know our Talawa community by visiting our YouTube channel. There you will find interviews, technical talks, feature reviews with volunteers and much more
+  <br/><br/>
+  You can follow the latest news from the Palisadoes team on Twitter.
+    `,
+  };
+
